Initial commit.
This should be entirely working now - MELONS!

Work as expected. A seed grows to a plant, plant flowers and an
adjacent melon block grows, and is attached with a stem. Once the
melon block is harvested (which drops 3-5 melon slices), the
stem is lost and the plant will grow a new melon from it's mature
stage.