allow filling the watercan with e.g. river water.
I didn't merge the part that would reject filling a watercan in
areas you are protected from editing. It would make more sense
if this was allowed, since the water block used to fill is not
consumed.

Also, I'd rather see code making sure that other people's plants
can't be overwatered, which is more dangereous and I've not checked
if that is possible now.