Pumpkins!
Since it's almost thanksgiving, here's some pumpkins, and
as usual, with a twist. In itself the pumpkins behave
like you expect - digging them yields a block, not slices.
You can get seeds from a block, or cook a block into roast
pumpkin, which is edible.

The catch is that there's only one piece of food from a pumpkin
so it's a lot of work. And if you want to expand your pumpkin
farm, you'll have to water as otherwise the plants die and
will not spawn more than 1 pumpkin if you're unlucky, so
it's going to be difficult and labor intensive to grow a
large pumpkin farm.

So yeah, maybe just grow pumpkins once a year for some
holiday cheer, like some countries do, and then back to
taters, cuz they easy.