Allow Disabling of Hydration mechanics
Disables the hydration mechanics when the user sets the game difficulty
to easy in the crops_settings.txt file. This makes the mod much easier
to play on easy mode. This was acheived by only executing the parts
of the code that pertain to hydration when the hydration variable in
settings is set to true. Otherwise the code that does hydration is
never executed.